Master

+++ server/openstorefront/openstorefront-web/src/main/java/edu/usu/sdl/openstorefront/web/test/user/UserWatchTest.java
@@ -19,7 +19,6 @@ import edu.usu.sdl.openstorefront.common.util.TimeUtil;
 import edu.usu.sdl.openstorefront.core.entity.UserWatch;
 import edu.usu.sdl.openstorefront.core.model.ComponentAll;
 import edu.usu.sdl.openstorefront.web.test.BaseTestCase;
-import edu.usu.sdl.openstorefront.web.test.component.ComponentTest;
 
 /**
  *
@@ -29,19 +28,22 @@ public class UserWatchTest
 		extends BaseTestCase
 {
 
-	public UserWatchTest()
+	private UserWatch userWatch = null;
+	ComponentAll componentAll = null;
+
+	@Override
+	protected void initializeTest()
 	{
-		this.description = "User Watch";
+		super.initializeTest();
+		results.append("Create component").append("<br>");
+		componentAll = getTestComponent();
 	}
 
 	@Override
 	protected void runInternalTest()
 	{
-		results.append("Create component").append("<br>");
-		ComponentAll componentAll = ComponentTest.createTestComponent();
-
 		results.append("Create watch").append("<br>");
-		UserWatch userWatch = new UserWatch();
+		userWatch = new UserWatch();
 		userWatch.setUsername(TEST_USER);
 		userWatch.setComponentId(componentAll.getComponent().getComponentId());
 		userWatch.setNotifyFlg(true);
@@ -55,12 +57,22 @@ public class UserWatchTest
 		results.append("Read watch").append("<br>");
 		userWatch = service.getUserService().getWatch(userWatch.getUserWatchId());
 
-		//remove
-		results.append("Remove watch").append("<br>");
-		service.getUserService().deleteWatch(userWatch.getUserWatchId());
+	}
 
-		results.append("Remove component").append("<br>");
-		ComponentTest.deleteComponent(componentAll.getComponent().getComponentId());
+	@Override
+	public String getDescription()
+	{
+		return "User Watch";
+	}
+
+	@Override
+	protected void cleanupTest()
+	{
+		super.cleanupTest();
+		if (userWatch != null) {
+			results.append("Remove watch").append("<br>");
+			service.getUserService().deleteWatch(userWatch.getUserWatchId());
+		}
 	}
 
 }
